# cmd/api/server/openapi.yaml
openapi: 3.1.0
info:
  title: Code Search API
  description: |
    A self-hosted code search and replace platform with support for batch operations across repositories.

    ## Features
    - Fast indexed search across all repositories (powered by Zoekt)
    - CLI for bulk operations (search & replace with MR creation)
    - GitHub & GitLab integration (clone, sync, create PRs/MRs)

    ## Authentication
    API requests can be authenticated using a Bearer token in the Authorization header.
  version: 1.0.0
  contact:
    name: Code Search Team
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://codesearch.example.com
    description: Production server
...
paths:
...
  /api/v1/repos/by-id/{id}/tree:
    get:
      tags:
        - File Browsing
      summary: List directory contents
      description: |
        Returns the contents of a directory in a repository.
        Can be disabled via `CS_UI_DISABLE_BROWSE_API`.
      operationId: listTree
      parameters:
        - $ref: "#/components/parameters/RepoId"
        - name: path
          in: query
          description: Path within the repository (empty for root)
          schema:
            type: string
            default: ""
        - name: ref
          in: query
          description: Git ref (branch, tag, or commit). Defaults to repository's default branch.
          schema:
            type: string
      responses:
        "200":
          description: Directory listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TreeResponse"
        "404":
          description: Repository or path not found
        "503":
          description: Repository not cloned locally (indexer unavailable in sharded mode)
  /api/v1/repos/by-id/{id}/blob:
    get:
      tags:
        - File Browsing
      summary: Get file contents
      description: |
        Returns the contents of a file in a repository.
        Can be disabled via `CS_UI_DISABLE_BROWSE_API`.
      operationId: getBlob
      parameters:
        - $ref: "#/components/parameters/RepoId"
        - name: path
          in: query
          required: true
          description: Path to the file within the repository
          schema:
            type: string
        - name: ref
          in: query
          description: Git ref (branch, tag, or commit). Defaults to repository's default branch.
          schema:
            type: string
      responses:
        "200":
          description: File contents
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlobResponse"
        "404":
          description: Repository or file not found
        "503":
          description: Repository not cloned locally
  /api/v1/repos/by-id/{id}/raw:
    get:
      tags:
        - File Browsing
      summary: Get raw file contents
      description: |
        Returns the raw contents of a file without JSON wrapping.
        Useful for downloading files directly.
        Can be disabled via `CS_UI_DISABLE_BROWSE_API`.
      operationId: getRawBlob
      parameters:
        - $ref: "#/components/parameters/RepoId"
        - name: path
          in: query
          required: true
          description: Path to the file within the repository
          schema:
            type: string
        - name: ref
          in: query
          description: Git ref (branch, tag, or commit). Defaults to repository's default branch.
          schema:
            type: string
      responses:
        "200":
          description: Raw file contents
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
            text/plain:
              schema:
                type: string
        "404":
          description: Repository or file not found
        "503":
          description: Repository not cloned locally
  /api/v1/repos/by-id/{id}/refs:
    get:
      tags:
        - File Browsing
      summary: List branches and tags
      description: |
        Returns all branches and tags for a repository.
        Can be disabled via `CS_UI_DISABLE_BROWSE_API`.
      operationId: getBranchesAndTags
      parameters:
        - $ref: "#/components/parameters/RepoId"
      responses:
        "200":
          description: Branches and tags
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefsResponse"
        "404":
          description: Repository not found or not cloned
  /api/v1/repos/by-id/{id}/symbols:
    get:
      tags:
        - File Browsing
      summary: Get file symbols
      description: |
        Returns symbols (functions, classes, methods) in a specific file using Tree-sitter.
        Can be disabled via `CS_UI_DISABLE_BROWSE_API`.
      operationId: getFileSymbols
      parameters:
        - $ref: "#/components/parameters/RepoId"
        - name: path
          in: query
          required: true
          description: Path to the file within the repository
          schema:
            type: string
        - name: ref
          in: query
          description: Git ref (branch, tag, or commit). Defaults to repository's default branch.
          schema:
            type: string
      responses:
        "200":
          description: File symbols
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileSymbolsResponse"
        "404":
          description: Repository or file not found
...
  /api/v1/scip/repos/{id}/files:
    get:
      tags:
        - SCIP
      summary: List indexed files
      description: List all files that are included in the SCIP index for a repository.
      operationId: listSCIPFiles
      parameters:
        - $ref: "#/components/parameters/RepoId"
        - name: limit
          in: query
          description: Maximum number of files to return
          schema:
            type: integer
            default: 100
            minimum: 1
            maximum: 1000
      responses:
        "200":
          description: List of indexed files
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListSCIPFilesResponse"
        "404":
          description: Repository not found

  # Scheduler Endpoints
...
components:
...
  parameters:
    RepoId:
      name: id
      in: path
      required: true
      description: Repository ID
      schema:
        type: integer
        format: int64
...
  schemas:
...
    TreeEntry:
      type: object
      properties:
        name:
          type: string
          description: File or directory name
          example: src
        path:
          type: string
          description: Full path within the repository
          example: src
        type:
          type: string
          enum: [file, dir]
          description: Entry type
          example: dir
        size:
          type: integer
          format: int64
          description: File size in bytes (only for files)
          example: 1234
    TreeResponse:
      type: object
      properties:
        entries:
          type: array
          items:
            $ref: "#/components/schemas/TreeEntry"
        path:
          type: string
          description: Current directory path
          example: src
        ref:
          type: string
          description: Git ref used
          example: main
    BlobResponse:
      type: object
      properties:
        content:
          type: string
          description: File content (empty for binary files)
        path:
          type: string
          description: File path
          example: src/main.go
        size:
          type: integer
          format: int64
          description: File size in bytes
          example: 1234
        language:
          type: string
          description: Detected programming language
          example: Go
        language_mode:
          type: string
          description: CodeMirror syntax highlighting mode
          example: go
        binary:
          type: boolean
          description: Whether the file is binary
          example: false
        ref:
          type: string
          description: Git ref used
          example: main
    RefsResponse:
      type: object
      properties:
        branches:
          type: array
          items:
            type: string
          example: [main, develop, feature/new-feature]
        tags:
          type: array
          items:
            type: string
          example: [v1.0.0, v1.1.0]
    FileSymbol:
      type: object
      properties:
        name:
          type: string
          description: Symbol name
          example: HandleRequest
        kind:
          type: string
          description: Symbol kind (function, class, method, variable, etc.)
          example: function
        line:
          type: integer
          description: Line number (1-based)
          example: 42
        column:
          type: integer
          description: Column number (0-based)
          example: 5
        signature:
          type: string
          description: Full signature (if available)
          example: "func HandleRequest(w http.ResponseWriter, r *http.Request)"
        parent:
          type: string
          description: Parent symbol name (for methods, nested types)
          example: Server
    FileSymbolsResponse:
      type: object
      properties:
        symbols:
          type: array
          items:
            $ref: "#/components/schemas/FileSymbol"
        path:
          type: string
          description: File path
          example: src/handler.go

    # Connection Schemas
...
    ListSCIPFilesResponse:
      type: object
      properties:
        files:
          type: array
          items:
            type: string
          description: List of file paths in the SCIP index
        count:
          type: integer
          description: Number of files returned

    # UI Settings Schema
...